tidy up setting of $GOROOT
This cleans out $GOROOT/bin from existing $PATH. This is mostly so when go is
upgraded, it doesn't keep the old path in there in tmux. In order to remove
the subpath we use the array representation $path to remove the matching
entry. Also unset $GOROOT after so it can be reset further down.  We don't do
anything with $GOPATH as it's not dependent on the go version.